# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/service_level_priorities:
    get:
      tags:
        - ServiceLevelPriorities
      summary: List ServiceLevelPriorities
      description: Get a paginated list of ServiceLevelPriorities with optional filtering and sorting
      operationId: listServiceLevelPriorities
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceLevelPriorityList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - ServiceLevelPriorities
      summary: Create ServiceLevelPriority
      description: Create a new ServiceLevelPriority
      operationId: createServiceLevelPriority
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceLevelPriorityRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceLevelPriority'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/warranty_claims:
    get:
      tags:
        - WarrantyClaims
      summary: List WarrantyClaims
      description: Get a paginated list of WarrantyClaims with optional filtering and sorting
      operationId: listWarrantyClaims
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarrantyClaimList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - WarrantyClaims
      summary: Create WarrantyClaim
      description: Create a new WarrantyClaim
      operationId: createWarrantyClaim
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWarrantyClaimRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarrantyClaim'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    ServiceLevelPriority:
      type: object
      required:
        - id
        - sla_id
        - priority
        - response_time_mins
        - resolution_time_mins
        - metadata
      properties:
        id:
          type: string
          format: uuid
        sla_id:
          type: string
          format: uuid
        priority:
          type: string
        response_time_mins:
          type: integer
          format: int32
        resolution_time_mins:
          type: integer
          format: int32
        metadata:
          type: object
...
    CreateServiceLevelPriorityRequest:
      type: object
      required:
        - sla_id
        - priority
        - response_time_mins
        - resolution_time_mins
        - metadata
      properties:
        sla_id:
          type: string
          format: uuid
        priority:
          type: string
        response_time_mins:
          type: integer
          format: int32
        resolution_time_mins:
          type: integer
          format: int32
        metadata:
          type: object
...
    ServiceLevelPriorityList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ServiceLevelPriority'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    WarrantyClaim:
      type: object
      required:
        - id
        - item_id
        - claim_date
        - is_under_warranty
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        customer_id:
          nullable: true
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        serial_no:
          nullable: true
          type: string
          maxLength: 140
        claim_date:
          type: string
          format: date-time
        warranty_expiry:
          nullable: true
          type: string
          format: date-time
        is_under_warranty:
          type: boolean
        status:
          type: string
        issue_id:
          nullable: true
          type: string
          format: uuid
        description:
          nullable: true
          type: string
          maxLength: 2000
        resolution:
          nullable: true
          type: string
          maxLength: 2000
        metadata:
          type: object
...
    CreateWarrantyClaimRequest:
      type: object
      required:
        - item_id
        - claim_date
        - is_under_warranty
        - status
        - metadata
      properties:
        customer_id:
          nullable: true
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        serial_no:
          nullable: true
          type: string
          maxLength: 140
        claim_date:
          type: string
          format: date-time
        warranty_expiry:
          nullable: true
          type: string
          format: date-time
        is_under_warranty:
          type: boolean
        status:
          type: string
        issue_id:
          nullable: true
          type: string
          format: uuid
        description:
          nullable: true
          type: string
          maxLength: 2000
        resolution:
          nullable: true
          type: string
          maxLength: 2000
        metadata:
          type: object
...
    WarrantyClaimList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/WarrantyClaim'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []